refactor: change ButtonAsIconWrapper to specifically a CrossIconButton as this follows the principles of atomic design more closely
