--- app/components/ui/atoms/ButtonAsIconWrapper.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-const ButtonAsIconWrapper = styled.button.attrs({
-  type: 'button',
-})`
-  background-color: transparent;
-  border: none;
-  line-height: 0;
-  padding: 0;
-`
-
-export default ButtonAsIconWrapper
